+++ llvm/lib/ObjCopy/ELF/ELFObject.cpp
@@ -2637,7 +2637,7 @@ template <class ELFT> Error ELFWriter<ELFT>::finalize() {
 Error BinaryWriter::write() {
   SmallVector<const SectionBase *, 30> LoadableSections;
   for (const SectionBase &Sec : Obj.allocSections()) {
-    if ((Sec.Flags & SectionFlag::SecLoad) && Sec.Type != SHT_NOBITS)
+    if (Sec.Type != SHT_NOBITS)
       LoadableSections.push_back(&Sec);
   }
 
@@ -2651,7 +2651,7 @@ Error BinaryWriter::write() {
 
   assert(LoadableSections.front()->Offset == 0);
 
-  for (unsigned i = 0; i != LoadableSections.size(); ++i) {
+  for (std::size_t i = 0; i != LoadableSections.size(); ++i) {
     const SectionBase &Sec = *LoadableSections[i];
     if (Error Err = Sec.accept(*SecWriter))
       return Err;
@@ -2661,9 +2661,9 @@ Error BinaryWriter::write() {
                              ? LoadableSections[i + 1]->Offset
                              : Buf->getBufferSize();
     assert(PadOffset <= Buf->getBufferSize());
-    if (Sec.Offset + Sec.Size < PadOffset)
-      std::fill(Buf->getBufferStart() + Sec.Offset + Sec.Size,
-                Buf->getBufferStart() + PadOffset, GapFill);
+    assert(Sec.Offset + Sec.Size <= PadOffset);
+    std::fill(Buf->getBufferStart() + Sec.Offset + Sec.Size,
+              Buf->getBufferStart() + PadOffset, GapFill);
   }
 
   // TODO: Implement direct writing to the output stream (without intermediate
@@ -2690,7 +2690,7 @@ Error BinaryWriter::finalize() {
   // file size. This might not be the same as the offset returned by
   // layoutSections, because we want to truncate the last segment to the end of
   // its last non-empty section, to match GNU objcopy's behaviour.
-  TotalSize = PadTo && PadTo.value() > MinAddr ? PadTo.value() - MinAddr : 0;
+  TotalSize = PadTo.value() > MinAddr ? PadTo.value() - MinAddr : 0;
   for (SectionBase &Sec : Obj.allocSections())
     if (Sec.Type != SHT_NOBITS && Sec.Size > 0) {
       Sec.Offset = Sec.Addr - MinAddr;